refactor(content-field): added a content-field interface
