+++ src/lib/models/Params.ts
@@ -1,0 +1,4 @@
+export interface Params {
+  instance: object;
+  installation: object;
+}
